requirements.txt
pillow
numpy
matplotlib
pyyaml
pytest
contextily
